RFCT: add context to allow state to be reused, and keep components clean from states and hooks.
