+++ frontend/src/features/settings-finder/subfeatures/upload-form/UploadForm.jsx
@@ -1,50 +1,21 @@
-import { useQueryClient } from "@tanstack/react-query";
-
 import { Form } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import CustomFormField from "@/components/widgets/custom-form-field/CustomFormField";
 import HoverButton from "@/components/widgets/hover-button/HoverButton";
 import { cn } from "@/lib/utils";
-import useSettingsStore from "@/store/settings";
 import useUploadFormValidate from "./hooks/useUploadFormValidate";
-import useSettingsFinder from "../../hooks/useSettingsFinder";
-import { useSettingsFinderContext } from "../../contexts/SettingsFinderContext";
-import showUploadToast from "./components/showUploadToast";
+import useUploadForm from "./hooks/useUploadForm";
 
 const UploadForm = ({ className, mode }) => {
-  const queryClient = useQueryClient();
-  const { item } = useSettingsStore();
-  const { setError, setImage } = useSettingsFinderContext();
-
   const {
-    action: { handleImageProcess },
-  } = useSettingsFinder({ setError });
+    state: { item },
+    action: { onFileChange },
+  } = useUploadForm();
 
   const {
     state: { ref, uploadFormInfo },
     action: { onSubmit, uploadForm },
   } = useUploadFormValidate();
-
-  const onFileChange = async (e) => {
-    e.preventDefault();
-    setError("");
-
-    const file = e.target.files[0];
-    if (file) {
-      queryClient.removeQueries();
-
-      const fileName = file.name;
-      const targetCount = uploadForm.getValues("quantity");
-
-      setImage(URL.createObjectURL(file));
-
-      showUploadToast({ mode, item, fileName, targetCount });
-
-      await handleImageProcess(mode, item, targetCount, file);
-
-      e.target.value = null;
-    }
-  };
 
   return (
     <div className={cn("h-full w-full", className)}>
@@ -79,7 +50,9 @@ const UploadForm = ({ className, mode }) => {
         type="file"
         accept=".png, .jpg"
         ref={ref}
-        onChange={onFileChange}
+        onChange={(e) =>
+          onFileChange(e, mode, uploadForm.getValues("quantity"))
+        }
       />
     </div>
   );
